requirements.txt
streamlit>=1.28.0
requests>=2.31.0
google-genai
python-dotenv>=1.0.0
langgraph>=1.0.0
langchain-google-genai>=3.0.0
langchain-core>=1.0.0
aiohttp>=3.11.0
